細かい修正

+++ app/controllers/codes_controller.rb
@@ -48,20 +48,22 @@ class CodesController < ApplicationController
     @comment = Comment.new
     @comments = @code.comments.includes(:user)
 
-    new_history = @code.histories.new
-    new_history.user_id = current_user.id
-
-    if current_user.histories.exists?(code_id: "#{params[:id]}")
-      old_history = current_user.histories.find_by(code_id: "#{params[:id]}")
-      old_history.destroy
-    end
-
-    new_history.save
-
-    histories_stock_limit = 10
-    histories = current_user.histories.all 
-    if histories.count > histories_stock_limit
-      histories[0].destroy
+    if user_signed_in?
+      new_history = @code.histories.new
+      new_history.user_id = current_user.id
+  
+      if current_user.histories.exists?(code_id: "#{params[:id]}")
+        old_history = current_user.histories.find_by(code_id: "#{params[:id]}")
+        old_history.destroy
+      end
+  
+      new_history.save
+  
+      histories_stock_limit = 10
+      histories = current_user.histories.all 
+      if histories.count > histories_stock_limit
+        histories[0].destroy
+      end
     end
 
   end
